requirements.txt
click>=8.0.0
gitpython>=3.1.0
pandas>=1.3.0
matplotlib>=3.4.0
plotly>=5.3.0
rich>=10.0.0
pytest>=7.0.0
